ref: refactor MainConsole to JobAdvertisement and ResumeUI
